Organize DTOs by item and list
