+++ TodoApi.Tests/Controllers/TodoListsControllerTests.cs
@@ -3,6 +3,7 @@ using Microsoft.EntityFrameworkCore;
 using TodoApi.Controllers;
 using TodoApi.Models;
 using TodoApi.Repositories;
+using TodoApi.Dtos.TodoLists;
 
 namespace TodoApi.Tests;
 
@@ -69,7 +70,7 @@ public class TodoListsControllerTests
 
             var result = await controller.PutTodoList(
                 3,
-                new Dtos.UpdateTodoList { Name = "Task 3" }
+                new UpdateTodoList { Name = "Task 3" }
             );
 
             Assert.IsType<NotFoundResult>(result);
@@ -89,7 +90,7 @@ public class TodoListsControllerTests
             var todoList = await context.TodoList.Where(x => x.Id == 2).FirstAsync();
             var result = await controller.PutTodoList(
                 todoList.Id,
-                new Dtos.UpdateTodoList { Name = "Changed Task 2" }
+                new UpdateTodoList { Name = "Changed Task 2" }
             );
 
             Assert.IsType<OkObjectResult>(result);
@@ -106,7 +107,7 @@ public class TodoListsControllerTests
             var repository = new TodoListRepository(context);
             var controller = new TodoListsController(repository);
 
-            var result = await controller.PostTodoList(new Dtos.CreateTodoList { Name = "Task 3" });
+            var result = await controller.PostTodoList(new CreateTodoList { Name = "Task 3" });
 
             Assert.IsType<CreatedAtActionResult>(result.Result);
             Assert.Equal(3, context.TodoList.Count());
