+++ TodoApi/Dtos/TodoItems/CreateTodoItem.cs
@@ -1,6 +1,6 @@
-namespace TodoApi.Dtos;
+namespace TodoApi.Dtos.TodoItems;
 
-public class TodoItemDto
+public class CreateTodoItem
 {
     public required string Description { get; set; }
     public bool IsCompleted { get; set; }
