+++ TodoApi/Dtos/TodoItems/TodoItemDto.cs
@@ -1,6 +1,6 @@
-namespace TodoApi.Dtos;
+namespace TodoApi.Dtos.TodoItems;
 
-public class CreateTodoItem
+public class TodoItemDto
 {
     public required string Description { get; set; }
     public bool IsCompleted { get; set; }
